Using the homebrew version of LibRAW

+++ build-osx.sh
@@ -1,7 +1,7 @@
 #! /bin/bash
 
 brew cask uninstall oclint
-brew reinstall little-cms2 fftw curl zlib exiv2 || exit 1
+brew reinstall little-cms2 fftw curl zlib exiv2 libraw || exit 1
 
 #HASH=9ba3d6ef8891e5c15dbdc9333f857b13711d4e97 #qt@5.5
 #QTPREFIX="qt@5.5"
@@ -19,13 +19,15 @@ export LD_LIBRARY_PATH="/usr/local/opt/curl/lib:/usr/local/opt/zlib/lib:/usr/loc
 mkdir -p hdrmerge/build || exit 1
 cd hdrmerge/build || exit 1
 
-rm -rf LibRaw
-git clone https://github.com/LibRaw/LibRaw.git || exit 1
-cd LibRaw || exit 1
-autoreconf --install || exit 1
-./configure --prefix=/usr/local || exit 1
-make -j2 install || exit 1
-cd ..
+if [ "x" = "y" ]; then
+	rm -rf LibRaw
+	git clone https://github.com/LibRaw/LibRaw.git || exit 1
+	cd LibRaw || exit 1
+	autoreconf --install || exit 1
+	./configure --prefix=/usr/local || exit 1
+	make -j2 install || exit 1
+	cd ..
+fi
 pwd
 curl -L http://www.alglib.net/translator/re/alglib-3.14.0.cpp.gpl.tgz -O || exit 1
 tar xf alglib-3.14.0.cpp.gpl.tgz || exit 1
